changed PlantPageData prices from double to long
*also changed RequestParam targetedPrice in PlantPageController to long
*added method centsToEuro to PriceHelper
*adjusted affected classes and tests

+++ user/src/main/java/org/dicadeveloper/weplantaforest/planting/PlantItem.java
@@ -11,6 +11,6 @@ public class PlantItem {
 
     int amount;
 
-    double treePrice;
+    long treePrice;
 
 }
